Create SelectAgeFragment and SelectGenderFragment
